requirements.txt
openai==1.54.4
pydantic==2.9.2
python-dotenv==1.0.0
tqdm==4.66.1
tree-sitter==0.23.2
tree-sitter-c==0.23.1
tree-sitter-cpp==0.23.4
